requirements.txt
flask==3.0.0
requests>=2.32.3
google-generativeai>=0.8.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dateutil==2.8.2
pytz==2023.3
apscheduler==3.10.4
redis==5.0.1
rq==1.15.1
cryptography>=3.4.8
ivrit==0.0.4